+++ CHANGES.md
@@ -1,6 +1,8 @@
 # 1.0.1
 
 * Treat `:js` requests as non-document, too.
+* `Controller#form` now returns the form object and not the operation.
+* In `Controller`, `#form`, `#present`, `#run` and `#respond` now all have the same API: `run(constant, options)`. If you want to pass a custom params hash, use `run Comment::Create, params: {..}`.
 
 # 1.0.0